Text Game(Journey to the Centre of the Earth)
This is a science fiction adventure film, that follows the story of a professor, his nephew, and their guide as they embark on an extraordinary adventure. The movie involves the trio discovering a mysterious book belonging to the professor's brother, which contains clues about a passage to the center of the Earth marking the first stage of the game. Driven by curiosity and the desire to uncover the truth about his brother's disappearance years ago, The professor and his companions embark on a perilous journey to follow the clues outlined in the book. The second stage of the game encounters the professor and his team descending in to the unknown abyss where they encounter numerous challenges, including treacherous landscapes, prehistoric creatures, and natural hazards as they navigate their way deeper into the Earth's core. Throughout the journey, they rely on teamwork, quick thinking, and courage to overcome obstacles and survive in the unfamiliar and dangerous underground world which marks the 3rd stage of the game. The adventure tests their resilience and leads them to astonishing discoveries about the Earth's hidden secrets and the possibility of finding a way back to the surface which is the last stage of the game. This is the link to the movie, IMDB LINK https://www.imdb.com/title/tt0373051/?ref_=ext_shr_lnk. 

##Dive into the world of coding by playing this  Game